refactor(idpe-17789): rename partitions_source mod to compaction_jobs_source mod
